Format code with forge fmt

+++ src/PriceConverter.sol
@@ -5,18 +5,13 @@ pragma solidity ^0.8.18;
 import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/shared/interfaces/AggregatorV3Interface.sol";
 
 library PriceConverter {
-    function getPrice(
-        AggregatorV3Interface priceFeed
-    ) internal view returns (uint256) {
-        (, int256 price, , , ) = priceFeed.latestRoundData();
+    function getPrice(AggregatorV3Interface priceFeed) internal view returns (uint256) {
+        (, int256 price,,,) = priceFeed.latestRoundData();
         // ETH/USD rate n 18 digits
-        return uint(price * 1e10);
+        return uint256(price * 1e10);
     }
 
-    function getConversionRate(
-        uint256 ethAmount,
-        AggregatorV3Interface priceFeed
-    ) internal view returns (uint256) {
+    function getConversionRate(uint256 ethAmount, AggregatorV3Interface priceFeed) internal view returns (uint256) {
         uint256 ethPrice = getPrice(priceFeed);
 
         uint256 ethAmountInUsd = (ethPrice * ethAmount) / 1e18;
@@ -26,8 +21,6 @@ library PriceConverter {
 
     function getVersion() internal view returns (uint256) {
         // 0x694AA1769357215DE4FAC081bf1f309aDC325306
-        return
-            AggregatorV3Interface(0x694AA1769357215DE4FAC081bf1f309aDC325306)
-                .version();
+        return AggregatorV3Interface(0x694AA1769357215DE4FAC081bf1f309aDC325306).version();
     }
 }
